requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
joblib==0.15.1
numpy==1.18.5
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
